Refactoring: add modules for webpack
