+++ momentum/src/index.html
@@ -24,7 +24,7 @@
                 <div class="song"><img src="../public/play.svg" alt="play" class="play-small">Summer Wind</div>
             </div>
         </div>
-        <div class="wheather">
+        <div class="weather">
             <div class="country">Minsk
                 <hr class="line">
                 </hr>
