+++ momentum/src/styles.css
@@ -2,9 +2,11 @@ body {
     padding: 15px;
     margin: 0;
     color: white;
-    background: url(./bg.jpeg) center/cover, rgba(0, 0, 0, 0.5);
+    background-image: url(./bg.jpeg);
+    background-position: center;
+    background-size: cover;
+    background-color: rgba(0, 0, 0, 0.5);
     background-blend-mode: multiply;
-    transition: background-image 1s ease-in-out;
     height: 100vh;
     display: flex;
     flex-direction: column;
